tools/requirements.txt
# MQTT测试程序依赖
paho-mqtt>=1.6.1
colorama>=0.4.4
